backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.5.0
httpx>=0.25.0

